move sortedIndex out of observable as other code will want to use it too
